chore(backend): rename wasm_backend module to backend
